hussh-client/requirements.txt
asn1crypto==0.24.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.2.2
idna==2.6
progressbar2==3.37.1
pycparser==2.18
python-utils==2.3.0
requests==2.18.4
six==1.11.0
urllib3==1.22
